Move properties to new indexing.IndexingMixin
